fixed the kernel's error, still have one CL_INVALID_KERNEL_NAME error but I don't know in which kernel I'm getting it

+++ rtpslib/system/flock/cl_src/rule_separation.cl
@@ -91,7 +91,7 @@ __kernel void rule_separation(
 
     // TODO: finish the computation of the separation
     if(pt.num_nearestFlockmates > 0){
-        pt.separation /= pt.num_nearestFlockmates;
+        pt.separation /= (float)pt.num_nearestFlockmates;
         pt.separation.w =0.f;
         pt.separation = normalize(pt.separation);
     }
